+++ pcs/lib/cib/constraint/ticket.py
@@ -12,6 +12,7 @@ from pcs.common import (
     reports,
 )
 from pcs.common.reports.item import ReportItem
+from pcs.lib.booth.config_validators import validate_ticket_name
 from pcs.lib.cib.constraint import constraint
 from pcs.lib.cib import tools
 from pcs.lib.errors import LibraryError
@@ -71,6 +72,8 @@ def prepare_options_with_set(cib, options, resource_set_list):
                 reports.messages.RequiredOptionsAreMissing(["ticket"])
             )
         )
+    else:
+        report_list.extend(validate_ticket_name(options["ticket"]))
     if report_list:
         raise LibraryError(*report_list)
     return options
@@ -87,6 +90,8 @@ def prepare_options_plain(cib, options, ticket, resource_id):
                 reports.messages.RequiredOptionsAreMissing(["ticket"])
             )
         )
+    else:
+        report_list.extend(validate_ticket_name(ticket))
     options["ticket"] = ticket
 
     if not resource_id: